requirements.txt
prometheus_client==0.16.0
python-dotenv==1.0.0
redis==4.5.4
selenium==4.8.3
webdriver_manager==3.8.6
